requirements.txt
contourpy==1.3.1
cycler==0.12.1
easyocr==1.7.2
filelock==3.18.0
fonttools==4.56.0
fsspec==2025.3.2
imageio==2.37.0
Jinja2==3.1.6
kiwisolver==1.4.8
lazy_loader==0.4
MarkupSafe==3.0.2
matplotlib==3.10.1
MouseInfo==0.1.3
mpmath==1.3.0
networkx==3.4.2
ninja==1.11.1.4
numpy==2.2.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
packaging==24.2
pillow==11.1.0
PyAutoGUI==0.9.54
pyclipper==1.3.0.post6
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyparsing==3.2.3
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python3-xlib==0.15
pytweening==1.2.0
PyYAML==6.0.2
scikit-image==0.25.2
scipy==1.15.2
shapely==2.0.7
six==1.17.0
sympy==1.13.1
tifffile==2025.3.30
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tqdm==4.67.1
triton==3.2.0
typing_extensions==4.13.0
